requirements.txt
appnope==0.1.3
attrs==22.1.0
backcall==0.2.0
bleach==3.3.1
certifi==2019.6.16
chardet==3.0.4
decorator==5.1.1
defusedxml==0.7.1
Django==2.2.2
django-adminplus==0.5
django-common==0.1.51
django-common-helpers==0.9.2
django-cron==0.5.1
django-dbbackup==3.2.0
django-extensions==3.1.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-storages==1.7.1
dropbox==9.4.0
entrypoints==0.3
idna==2.8
importlib-metadata==2.1.3
ipykernel==5.5.6
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
mysqlclient==1.4.2.post1
nbconvert==5.6.1
nbformat==5.1.3
notebook==5.7.16
packaging==20.9
pandocfilters==1.5.1
parso==0.7.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.12.0
prompt-toolkit==1.0.18
ptyprocess==0.7.0
Pygments==2.11.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.9.0.post0
pytz==2019.1
pyzmq==20.0.0
requests==2.22.0
Send2Trash==1.8.2
simplegeneric==0.8.1
six==1.12.0
sqlparse==0.3.0
terminado==0.8.3
testpath==0.6.0
tornado==6.1
traitlets==4.3.3
urllib3==1.25.3
wcwidth==0.2.13
webencodings==0.5.1
zipp==1.2.0
